refactor(server): moved message handlers in their own classes
